Move managed-api out of transpilation module
