+++ src/index.ts
@@ -6,5 +6,6 @@ export * from "./CompilerOptions";
 export * from "./LuaAST";
 export { LuaLibFeature } from "./LuaLib";
 export * from "./LuaPrinter";
+export * from "./managed-api";
 export * from "./transformation/context";
 export * from "./transpilation";
