+++ src/transpilation/index.ts
@@ -1,5 +1,4 @@
 export { Chunk } from "./chunk";
-export * from "./managed-api";
 export { Module } from "./module";
 export { Plugin } from "./plugins";
 export * from "./transpile";
